Converted some components to c++ classes, consolidated isr and irq calls into a arch based class.
