+++ src/drivers/keyboard.h
@@ -1,6 +1,8 @@
 #ifndef KEYBOARD_H
 #define KEYBOARD_H
 
+#include "../kernel/arch/x86/x86.h"
+
 typedef struct {
   bool shift;
   bool ctrl;
@@ -15,7 +17,7 @@ void update_flags(unsigned char code);
 unsigned int get_keyboard_pointer();
 void keyboard_push(unsigned char data);
 unsigned char keyboard_pull();
-void keyboard_install();
+void keyboard_install(x86* sys);
 unsigned char get_char(int keycode);
 
 #endif
