+++ src/kernel/arch/x86/interrupts.h
@@ -1,0 +1,84 @@
+#ifndef INTTERUPTS_H
+#define INTTERUPTS_H
+
+extern "C" void irq0();
+extern "C" void irq1();
+extern "C" void irq2();
+extern "C" void irq3();
+extern "C" void irq4();
+extern "C" void irq5();
+extern "C" void irq6();
+extern "C" void irq7();
+extern "C" void irq8();
+extern "C" void irq9();
+extern "C" void irq10();
+extern "C" void irq11();
+extern "C" void irq12();
+extern "C" void irq13();
+extern "C" void irq14();
+extern "C" void irq15();
+
+extern "C" void isr0();
+extern "C" void isr1();
+extern "C" void isr2();
+extern "C" void isr3();
+extern "C" void isr4();
+extern "C" void isr5();
+extern "C" void isr6();
+extern "C" void isr7();
+extern "C" void isr8();
+extern "C" void isr9();
+extern "C" void isr10();
+extern "C" void isr11();
+extern "C" void isr12();
+extern "C" void isr13();
+extern "C" void isr14();
+extern "C" void isr15();
+extern "C" void isr16();
+extern "C" void isr17();
+extern "C" void isr18();
+extern "C" void isr19();
+extern "C" void isr20();
+extern "C" void isr21();
+extern "C" void isr22();
+extern "C" void isr23();
+extern "C" void isr24();
+extern "C" void isr25();
+extern "C" void isr26();
+extern "C" void isr27();
+extern "C" void isr28();
+extern "C" void isr29();
+extern "C" void isr30();
+extern "C" void isr31();
+
+extern "C" void irq_handler(struct regs *r);
+
+void( *irq_routines[16])(struct regs*) = {
+  0, 0, 0, 0, 0, 0, 0, 0,
+  0, 0, 0, 0, 0, 0, 0, 0
+};
+
+const char *exception_messages[] = {
+  "Division by Zero",
+  "Debug Exception",
+  "Non Maskable Interrupt Exception",
+  "Breakpoint Exception",
+  "Into Detected Overflow Exception",
+  "Out of Bounds Exception",
+  "Invalid Opcode Exception",
+  "No Coprocessor Exception",
+  "Double Fault Exception",
+  "Coprocessor Segment Overrun Exception",
+  "Bad TSS Exception",
+  "Segment Not Present Exception",
+  "Stack Fault Exception",
+  "General Protection Fault Exception",
+  "Page Fault Exception",
+  "Unknown Interrupt Exception",
+  "Coprocessor Fault Exception",
+  "Alignment Check Exception (486+)",
+  "Machine Check Exception (Pentium/586+)",
+  "Reserved"
+};
+
+#endif
