--- src/kernel/arch/x86/irq.cpp
@@ -1,66 +1,0 @@
-#include "irq.h"
-#include "x86.h"
-#include "system.h"
-#include "../../../drivers/io.h"
-
-void( *irq_routines[16])(struct regs*) = {
-  0, 0, 0, 0, 0, 0, 0, 0,
-  0, 0, 0, 0, 0, 0, 0, 0
-};
-
-void irq_install_handler(int irq, void (*handler)(struct regs *r)) {
-  irq_routines[irq] = handler;
-}
-
-void irq_uninstall_handler(int irq) {
-  irq_routines[irq] = 0;
-}
-
-void irq_remap(void) {
-  io.outportb(0x20, 0x11);
-  io.outportb(0xA0, 0x11);
-  io.outportb(0x21, 0x20);
-  io.outportb(0xA1, 0x28);
-  io.outportb(0x21, 0x04);
-  io.outportb(0xA1, 0x02);
-  io.outportb(0x21, 0x01);
-  io.outportb(0xA1, 0x01);
-  io.outportb(0x21, 0x0);
-  io.outportb(0xA1, 0x0);
-}
-
-void irq_install() {
-  irq_remap();
-
-  idt_set_gate(32, (unsigned)irq0, 0x08, 0x8E);
-  idt_set_gate(33, (unsigned)irq1, 0x08, 0x8E);
-  idt_set_gate(34, (unsigned)irq2, 0x08, 0x8E);
-  idt_set_gate(35, (unsigned)irq3, 0x08, 0x8E);
-  idt_set_gate(36, (unsigned)irq4, 0x08, 0x8E);
-  idt_set_gate(37, (unsigned)irq5, 0x08, 0x8E);
-  idt_set_gate(38, (unsigned)irq6, 0x08, 0x8E);
-  idt_set_gate(39, (unsigned)irq7, 0x08, 0x8E);
-  idt_set_gate(40, (unsigned)irq8, 0x08, 0x8E);
-  idt_set_gate(41, (unsigned)irq9, 0x08, 0x8E);
-  idt_set_gate(42, (unsigned)irq10, 0x08, 0x8E);
-  idt_set_gate(43, (unsigned)irq11, 0x08, 0x8E);
-  idt_set_gate(44, (unsigned)irq12, 0x08, 0x8E);
-  idt_set_gate(45, (unsigned)irq13, 0x08, 0x8E);
-  idt_set_gate(46, (unsigned)irq14, 0x08, 0x8E);
-  idt_set_gate(47, (unsigned)irq15, 0x08, 0x8E);
-}
-
-extern "C" void irq_handler(struct regs *r) {
-  void (*handler)(struct regs *r);
-
-  handler = irq_routines[r->int_no - 32];
-  if(handler) {
-    handler(r);
-  }
-
-  if(r->int_no >= 40) {
-    io.outportb(0xA0, 0x20);
-  }
-
-  io.outportb(0x20, 0x20);
-}
