+++ src/kernel/interrupts/timer.cpp
@@ -1,10 +1,9 @@
 #include "timer.h"
-#include "../arch/x86/irq.h"
 #include "../../drivers/io.h"
 
-int timer_ticks = 0;
+unsigned long timer_ticks = 0;
 
-void timer_phase(int hz) {
+void timer_phase() {
   int divisor = TIMER_CLOCK_SPEED;
   io.outportb(TIMER_COMMAND_REGISTER, 0x36);
   io.outportb(TIMER_CHANNEL_0, divisor & 0xFF);
@@ -15,8 +14,8 @@ void timer_handler(struct regs *r) {
   timer_ticks++;
 }
 
-void timer_install() {
-  irq_install_handler(0, timer_handler);
+void timer_install(x86* sys) {
+  sys->irq_install_handler(0, timer_handler);
 }
 
 void timer_wait(int ticks) {
